修改product表，添加过期字段；修改excel接口

+++ lhzs-service/src/main/java/cn/lhzs/service/intf/ProductService.java
@@ -21,7 +21,7 @@ public interface ProductService {
 
     List<Product> searchProduct(String key);
 
-    Integer getCount();
+    Long getCount();
 
     JSONObject getProds(String data);
 }
